fix tipe data durasi dan rating film serta minor update

+++ final_71190523/app/src/main/java/com/revyn/final_71190523/Film.kt
@@ -1,3 +1,3 @@
 package com.revyn.final_71190523
 
-data class Film (val film: String, val durasi: String, val genre: String, val rating: Int)
+data class Film (val film: String, val durasi: String, val genre: String, val rating: Double)
